app/requirements.txt
flask==2.3.2
web3==6.4.0
py-solc-x==2.0.3
python-dotenv==1.0.1
langchain==0.3.14
numpy==1.26.4
werkzeug==3.0.1
langchain-google-genai==2.0.1
pydantic==2.8.0
google-cloud-aiplatform==1.36.4
flask-cors==5.0.0
